Remove physics/{GFDL,NRL}_parse_tracers.F90 (#27)
This PR removes physics/{GFDL,NRL}_parse_tracers.F90 after adding it in the previous commit. This is based on a discussion between the CCPP physics code managers and developers. These two files are host-model specific and do not belong into the ccpp-physics repository. The agreement is that both files are removed from the ccpp-physics repo and added to the host model repo instead.

See https://github.com/NCAR/ccpp-physics/pull/1087 for more information, and NEPTUNE/neptune_atmos#819 for the corresponding PR in neptune_atmos.